// src/types/ICreditsTv.ts
export interface ICreditsTv {
    cast: ICastTv[];
    crew: ICrewTv[];
    id: number;
}

export interface ICastTv {
    adult: boolean;
    gender: number;
    id: number;
    known_for_department: string;
    name: string;
    original_name: string;
    popularity: number;
    profile_path: string;
    character: string;
    credit_id: string;
    order: number;
}

export interface ICrewTv {
    adult: boolean;
    gender: number;
    id: number;
    known_for_department: string;
    name: string;
    original_name: string;
    popularity: number;
    profile_path: string;
    credit_id: string;
    department: string;
    job: string;
}